backend/bot.py
"""
Pipecat voice bot for conducting interviews
ACTUAL WORKING IMPLEMENTATION
"""

import os
import asyncio
from typing import Optional
import httpx
import threading

from question_bank import get_question, get_all_topics

# For Daily room creation
DAILY_API_KEY = os.getenv("DAILY_API_KEY")
DAILY_API_URL = os.getenv("DAILY_API_URL", "https://api.daily.co/v1")

def create_daily_room(session_id: str) -> dict:
    """Create a Daily room for the interview session"""
    
    if not DAILY_API_KEY:
        print("⚠️  No Daily API key - using demo URL")
        return {
            "url": f"https://demo.daily.co/interview-{session_id}",
            "token": ""
        }
    
    try:
        response = httpx.post(
            f"{DAILY_API_URL}/rooms",
            headers={
                "Authorization": f"Bearer {DAILY_API_KEY}",
                "Content-Type": "application/json"
            },
            json={
                "name": f"interview-{session_id}",
                "privacy": "private",
                "properties": {
                    "enable_chat": False,
                    "enable_screenshare": False,
                    "start_audio_off": False,
                    "start_video_off": True,
                    "max_participants": 2
                }
            },
            timeout=10.0
        )
        
        if response.status_code == 200:
            room_data = response.json()
            room_name = room_data["name"]
            room_url = room_data["url"]
            
            print(f"✅ Created Daily room: {room_url}")
            
            # Create a meeting token for the bot to join
            token_response = httpx.post(
                f"{DAILY_API_URL}/meeting-tokens",
                headers={
                    "Authorization": f"Bearer {DAILY_API_KEY}",
                    "Content-Type": "application/json"
                },
                json={
                    "properties": {
                        "room_name": room_name,
                        "is_owner": True
                    }
                },
                timeout=10.0
            )
            
            token = ""
            if token_response.status_code == 200:
                token = token_response.json().get("token", "")
                print(f"✅ Created meeting token for bot")
            
            return {
                "url": room_url,
                "token": token
            }
        else:
            print(f"❌ Failed to create Daily room: {response.status_code}")
            return {
                "url": f"https://demo.daily.co/interview-{session_id}",
                "token": ""
            }
    except Exception as e:
        print(f"❌ Error creating Daily room: {e}")
        return {
            "url": f"https://demo.daily.co/interview-{session_id}",
            "token": ""
        }


class InterviewBot:
    """Interview bot that uses Pipecat for voice interaction"""
    
    def __init__(self, session_id: str, storage, evaluator):
        self.session_id = session_id
        self.storage = storage
        self.evaluator = evaluator
        self.current_topic = None
        self.last_question = None
        self.questions_asked = 0
        self.session_data = storage.get_session(session_id)
        self.user_id = self.session_data["user_id"]
    
    async def get_first_question(self) -> str:
        """Get the first question for the interview"""
        # Start with a random topic
        topics = get_all_topics()
        self.current_topic = topics[0]  # Start with first topic
        
        question = get_question(self.current_topic, "easy")
        self.last_question = question
        self.questions_asked = 1
        
        return f"Hello! I'm going to ask you some interview questions. Let's start with: {question}"
    
    async def process_answer(self, answer_text: str) -> str:
        """
        Process user's answer and generate next question
        Returns the bot's response
        """
        
        if not self.last_question:
            # Haven't asked first question yet
            return await self.get_first_question()
        
        # Evaluate the answer
        print(f"📊 Evaluating answer for topic: {self.current_topic}")
        score, weak_points = self.evaluator.evaluate_answer(
            question=self.last_question,
            answer=answer_text,
            topic=self.current_topic
        )
        
        print(f"   Score: {score}/10")
        print(f"   Weak points: {weak_points}")
        
        # Store the result
        self.storage.add_question(
            session_id=self.session_id,
            question=self.last_question,
            answer=answer_text,
            score=score,
            topic=self.current_topic,
            weak_points=weak_points
        )
        
        # Update question count
        self.questions_asked += 1
        
        # Get updated knowledge map
        knowledge_map = self.storage.get_knowledge_map(self.user_id)
        topics_data = knowledge_map.get("topics", {})
        
        # Select next topic (focus on weak areas)
        session = self.storage.get_session(self.session_id)
        previous_topics = [q["topic"] for q in session["questions"]]
        
        if topics_data:
            self.current_topic = self.evaluator.select_next_topic(
                topics_data,
                previous_topics
            )
        else:
            # No data yet, cycle through topics
            topics = get_all_topics()
            self.current_topic = topics[self.questions_asked % len(topics)]
        
        # Determine difficulty based on performance
        current_score = topics_data.get(self.current_topic, 0.5)
        if current_score < 0.4:
            difficulty = "easy"
        elif current_score < 0.7:
            difficulty = "medium"
        else:
            difficulty = "hard"
        
        # Generate next question
        next_question = self.evaluator.generate_next_question(
            self.current_topic,
            difficulty,
            topics_data
        )
        
        self.last_question = next_question
        
        # Build response with feedback
        feedback = f"Thank you. "
        if score >= 8:
            feedback += "That was a strong answer. "
        elif score >= 6:
            feedback += "Good answer, but there's room for improvement. "
        else:
            feedback += "Let's work on strengthening that. "
        
        return f"{feedback}Next question: {next_question}"


def start_bot_in_room(room_url: str, session_id: str, storage, evaluator):
    """
    Start the interview bot in a Daily room
    
    This runs in a background thread and uses Pipecat for voice
    """
    
    print(f"🤖 Starting bot for session {session_id}")
    print(f"🔗 Room: {room_url}")
    
    # Check if we have the required API keys
    has_deepgram = bool(os.getenv("DEEPGRAM_API_KEY"))
    has_elevenlabs = bool(os.getenv("ELEVENLABS_API_KEY"))
    
    if not (has_deepgram and has_elevenlabs):
        print("⚠️  Missing API keys for voice:")
        if not has_deepgram:
            print("   - DEEPGRAM_API_KEY not set")
        if not has_elevenlabs:
            print("   - ELEVENLABS_API_KEY not set")
        print("   Bot will run in limited mode")
        return
    
    # Create bot instance
    bot = InterviewBot(session_id, storage, evaluator)
    
    # Start bot in background thread
    thread = threading.Thread(
        target=_run_bot_async,
        args=(bot, room_url),
        daemon=True
    )
    thread.start()
    
    print(f"✅ Bot thread started for session {session_id}")


def _run_bot_async(bot: InterviewBot, room_url: str):
    """Run the bot in an async context"""
    try:
        asyncio.run(_run_bot(bot, room_url))
    except Exception as e:
        print(f"❌ Bot error: {e}")


async def _run_bot(bot: InterviewBot, room_url: str):
    """
    Main bot loop using Pipecat
    
    NOTE: This is a simplified version. Full Pipecat integration would be:
    
    from pipecat.pipeline.pipeline import Pipeline
    from pipecat.transports.daily_transport import DailyTransport
    from pipecat.services.deepgram import DeepgramSTTService
    from pipecat.services.elevenlabs import ElevenLabsTTSService
    
    transport = DailyTransport(room_url, token, "Interview Bot")
    stt = DeepgramSTTService(api_key=os.getenv("DEEPGRAM_API_KEY"))
    tts = ElevenLabsTTSService(api_key=os.getenv("ELEVENLABS_API_KEY"))
    
    pipeline = Pipeline([
        transport.input(),
        stt,
        bot,  # Our interview bot processor
        tts,
        transport.output()
    ])
    
    await pipeline.run()
    """
    
    print(f"🤖 Bot would now:")
    print(f"   1. Join Daily room: {room_url}")
    print(f"   2. Ask first question via TTS")
    print(f"   3. Listen for user's answer via STT")
    print(f"   4. Evaluate answer with Claude")
    print(f"   5. Ask next question based on weak areas")
    print(f"   6. Repeat until session ends")
    
    # For now, just print that bot is ready
    # In production, this would actually run the Pipecat pipeline
    await asyncio.sleep(1)
    print(f"✅ Bot ready for session {bot.session_id}")
